requirements.txt
aiohttp[speedups]
aiohttp_cors
requests
toml
Pillow
PyJWT